// src/app/entities/ugp-interface.ts
export interface UGPInterface {
  id?: number;
  nombre?: string;
  descripcion?: string;
  abreviado?: string;
  anio_inicio?: number;
  anio_fin?: number;
  id_estado?: number;
  nombre_estado?: string;
  fecha_registro?: string;
  id_apertura?: number;
  id_proceso?: number;
  codigo_proceso?: string;
  nombre_proceso?: string;
  id_responsable?: number;
  nombre_responsable?: string;
  codigo_organo?: string;
  usuario_registro?: string;
  nombre_usuario?: string;
  id_organo?: number;
  id_etapa ?: number;
  id_ciclo ?: number;
}